+++ src/SimpleMemory.cpp
@@ -63,7 +63,7 @@ Result SimpleMemory::Write(IMemoryCallback& callback, MemAddr address, void* dat
         request.data.tag  = tag;
         request.done      = 0;
         request.write     = true;
-        memcpy(request.data.data, data, size);
+        memcpy(request.data.data, data, (size_t)size);
 
         // Broadcast the snoop data
         for (set<IMemoryCallback*>::iterator p = m_caches.begin(); p != m_caches.end(); ++p)
